+++ src/components/Documentation/Layout/SearchForm/index.js
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react'
 import Promise from 'promise-polyfill'
-import { loadResource } from '../../utils/resources'
+import { loadResource } from '../../../../utils/resources'
 
 import { SearchArea, Input, Wrapper } from './styles'
