+++ src/components/HeroSection/index.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 
 import { Container, Wrapper } from './styles'
 
-export default function Hero({ children }) {
+export default function HeroSection({ children }) {
   return (
     <Wrapper>
       <Container>{children}</Container>
@@ -11,6 +11,6 @@ export default function Hero({ children }) {
   )
 }
 
-Hero.propTypes = {
+HeroSection.propTypes = {
   children: PropTypes.node.isRequired
 }
