+++ src/components/Home/Diagram/index.js
@@ -8,9 +8,9 @@ import Slider from 'react-slick'
 import 'slick-carousel/slick/slick.css'
 import 'slick-carousel/slick/slick-theme.css'
 
-import Link from '../Link'
+import Link from '../../Link'
 
-import { OnlyDesktop, OnlyMobile } from '../../styles'
+import { OnlyDesktop, OnlyMobile } from '../../../styles'
 
 import {
   Abstract,
