+++ src/components/Home/index.js
@@ -1,29 +1,29 @@
 import React from 'react'
 
-import LearnMore from '../LearnMore'
-import Hero from '../Hero'
-import LandingHero from '../LandingHero'
-import Diagram from '../Diagram'
-import PromoSection from '../PromoSection'
-import UseCases from '../UseCases'
-import Subscribe from '../Subscribe'
+import LearnMore from './LearnMore'
+import HeroSection from '../HeroSection'
+import LandingHero from './LandingHero'
+import Diagram from './Diagram'
+import PromoSection from './PromoSection'
+import UseCases from './UseCases'
+import SubscribeSection from '../SubscribeSection'
 
 import { LearnMoreSection } from './styles'
 
 export default function HomePage() {
   return (
     <>
-      <Hero>
+      <HeroSection>
         <LandingHero />
         <LearnMoreSection>
           <LearnMore />
         </LearnMoreSection>
         <span name="nextSlide" style={{ marginTop: `-58px` }} />
-      </Hero>
+      </HeroSection>
       <Diagram />
       <PromoSection />
       <UseCases />
-      <Subscribe />
+      <SubscribeSection />
     </>
   )
 }
