--- src/components/TextRotate/styles.js
@@ -1,7 +1,0 @@
-import styled from 'styled-components'
-
-export const Wrapper = styled.span``
-
-export const Cursor = styled.span`
-  vertical-align: 4px;
-`
